+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+- fix: fix bug in `Doki` class
+- fix: badge widget color luminance
+- docs: update readme and added informations about queries in AndroidManifest.xml
+
 ## 1.0.0+3
 - style: fix code style
 ## 1.0.0+2